push smoke_spinner to a repo so we can show people
this is a hastily-written modification of the original `smoke_tests.py`
which tests all possible 1-input 1-output z_sendmany transactions
It creates a single chain of transactions, where each transaction's
input is the output of the previous transaction. since it creates a loop
of transactions, it's called "the spinner".

this is incredibly slow, since it requires at least one block-time per
transaction type.